requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.135.1
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-core==1.2.16
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.9
numpy==2.2.6
openai==2.24.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
portalocker==3.2.0
protobuf==7.34.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==5.3.0
python-dotenv==1.2.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.17.0
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.47
starlette==0.52.1
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
